+++ src/main/java/fll/TournamentTeam.java
@@ -6,8 +6,14 @@
 
 package fll;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Iterator;
 import java.util.List;
+
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -69,6 +75,60 @@ public class TournamentTeam extends Team {
       final String eventDivision = t.getAwardGroup();
       if (!eventDivision.equals(divisionStr)) {
         iter.remove();
+      }
+    }
+  }
+
+  /**
+   * Builds a team object from its database info given the team number.
+   * 
+   * @param connection Database connection.
+   * @param teamNumber Number of the team for which to build an object.
+   * @param tournament the tournament to find the team at
+   * @return The new Team object or null if the team was not found in the
+   *         database.
+   * @throws SQLException on a database access error.
+   * @throws IllegalArgumentException if the team number is for {@link Team#NULL},
+   *           {@link Team#TIE} or {@link Team#BYE}
+   */
+  public static @Nullable TournamentTeam getTournamentTeamFromDatabase(final Connection connection,
+                                                                       final Tournament tournament,
+                                                                       final int teamNumber)
+      throws SQLException {
+
+    // First, handle known non-database team numbers...
+    if (teamNumber == NULL_TEAM_NUMBER) {
+      throw new IllegalArgumentException("Cannot get tournament team for null number");
+    }
+    if (teamNumber == TIE_TEAM_NUMBER) {
+      throw new IllegalArgumentException("Cannot get tournament team for tie number");
+    }
+    if (teamNumber == BYE_TEAM_NUMBER) {
+      throw new IllegalArgumentException("Cannot get tournament team for bye number");
+    }
+
+    try (PreparedStatement stmt = connection.prepareStatement("SELECT Teams.Organization"//
+        + ", Teams.TeamName"//
+        + ", TournamentTeams.event_division" //
+        + ", TournamentTeams.judging_station" //
+        + " FROM Teams, TournamentTeams" //
+        + " WHERE Teams.TeamNumber = TournamentTeams.TeamNumber"//
+        + " AND TournamentTeams.Tournament = ?" //
+        + " AND Teams.TeamNumber = ?")) {
+      stmt.setInt(1, tournament.getTournamentID());
+      stmt.setInt(2, teamNumber);
+      try (ResultSet rs = stmt.executeQuery()) {
+        if (rs.next()) {
+          final String org = rs.getString(1);
+          final String name = rs.getString(2);
+          final String awardGroup = rs.getString(3);
+          final String judgingGroup = rs.getString(4);
+
+          final TournamentTeam x = new TournamentTeam(teamNumber, org, name, awardGroup, judgingGroup);
+          return x;
+        } else {
+          return null;
+        }
       }
     }
   }